# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BMLT",
        "description": "BMLT Admin API Documentation",
        "license": {
            "name": "MIT",
            "url": "https://github.com/bmlt-enabled/bmlt-server/blob/main/LICENSE"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/auth/token": {
            "post": {
                "tags": [
                    "server"
                ],
                "summary": "Creates a token",
                "description": "Exchange credentials for a new token",
                "operationId": "authToken",
                "requestBody": {
                    "description": "User credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TokenCredentials"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Returns when POST is successful.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Token"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Returns when credentials are incorrect.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthenticationError"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Returns when unauthorized.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthorizationError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/refresh": {
            "post": {
                "tags": [
                    "server"
                ],
                "summary": "Revokes and issues a new token",
                "description": "Refresh token.",
                "operationId": "authRefresh",
                "responses": {
                    "200": {
                        "description": "Returns when refresh is successful.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Token"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Returns when request is unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthenticationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bmltToken": []
                    }
                ]
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "tags": [
                    "server"
                ],
                "summary": "Revokes a token",
                "description": "Revoke token and logout.",
                "operationId": "authLogout",
                "responses": {
                    "200": {
                        "description": "Returns when token was logged out."
                    },
                    "401": {
                        "description": "Returns when request is unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthenticationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bmltToken": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AuthenticationError": {
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
            "AuthorizationError": {
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "This action is unauthorized."
                    }
                },
                "type": "object"
            },
...
            "ValidationError": {
                "required": [
                    "message",
                    "errors"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The field is required. (and 1 more error)"
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "example": "error details"
                            }
                        }
                    }
                },
                "type": "object"
            },
...
            "Token": {
                "required": [
                    "access_token",
                    "expires_at",
                    "token_type",
                    "user_id"
                ],
                "properties": {
                    "access_token": {
                        "type": "string",
                        "example": "2|tR6PIqa8tiBJWMu4zyb3qw4eECuERjLd7xeLKgBu"
                    },
                    "expires_at": {
                        "type": "integer",
                        "example": "1667342171"
                    },
                    "token_type": {
                        "type": "string",
                        "example": "bearer"
                    },
                    "user_id": {
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
            "TokenCredentials": {
                "required": [
                    "username",
                    "password"
                ],
                "properties": {
                    "username": {
                        "type": "string",
                        "format": "username",
                        "example": "MyUsername"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "PassWord12345"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bmltToken": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "tokenUrl": "api/v1/auth/token",
                        "refreshUrl": "api/v1/auth/refresh",
                        "scopes": {}
                    }
                }
            }
        }
    },
...
}